connect command to ADC read
for getTemperature and B0 AG
ADC convert all input and conversion is slow, don't know why.

+++ userspace/hw_lock.h
@@ -28,9 +28,9 @@ int lock_write_traces(int32_t* b0_current_uamps, int32_t* gx_current_uamps);
 
 int lock_init_board();
 
-int lock_read_board_temperature();
+double lock_read_board_temperature();
 
-int lock_read_b0_art_ground_current(int dropCount, int numAvg);
+double lock_read_b0_art_ground_current(int dropCount, int numAvg);
